Fix 64 bit compilation problem
git-svn-id: file:///tmp/flactag-git-move/svnroot/branches/v1.1@210 c94211a6-d224-44d1-9852-295abbaa3a11

+++ flactag.cc
@@ -156,7 +156,7 @@ CFlacTag::CFlacTag(const CCommandLine& CommandLine)
 								AlbumNum=count;
 								MatchFound=true;
 								
-								printf("%s: Album ID in album %d is a match\n",m_FlacFile.c_str(),count);
+								printf("%s: Album ID in album %d is a match\n",m_FlacFile.c_str(),(int)count);
 							}
 						}
